backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
torch==2.2.1
transformers==4.38.1
mistralai==0.1.3
python-dotenv==1.0.1
nltk==3.8.1
gunicorn==21.2.0